requirements.txt
Fiona==1.5.1
Shapely==1.5.8
click==4.0
cligj==0.1.0
decorator==3.4.2
networkx==1.9.1
numpy==1.9.2
pandas==0.16.1
pyproj==1.9.4
python-dateutil==2.4.2
pytz==2015.4
scipy==0.15.1
six==1.9.0
wsgiref==0.1.2
